Added Hymns Content Utils
