+++ src/hazzat-api.njsproj
@@ -48,6 +48,7 @@
     <TypeScriptCompile Include="Models\IHymnInfo.ts" />
     <TypeScriptCompile Include="Models\IFormatInfo.ts" />
     <TypeScriptCompile Include="Models\IServiceInfo.ts" />
+    <TypeScriptCompile Include="Providers\ServiceProviders\Helpers\HazzatContentUrils.ts" />
     <TypeScriptCompile Include="Providers\ServiceProviders\HymnsServiceProvider.ts" />
     <TypeScriptCompile Include="Providers\ServiceProviders\IHymnsServiceProvider.ts" />
     <TypeScriptCompile Include="Routes\ResourceTypes.ts" />
@@ -94,6 +95,7 @@
     <Folder Include="Configs\" />
     <Folder Include="Providers\" />
     <Folder Include="Providers\ServiceProviders\" />
+    <Folder Include="Providers\ServiceProviders\Helpers\" />
     <Folder Include="test\" />
     <Folder Include="Routes\" />
     <Folder Include="test\Helpers\" />
